+++ services/service-wakeup-song/src/main/kotlin/com/b1nd/dodamdodam/wakeupsong/application/data/response/Mp3DownloadResponse.kt
@@ -1,0 +1,8 @@
+package com.b1nd.dodamdodam.wakeupsong.application.data.response
+
+import java.nio.file.Path
+
+data class Mp3DownloadResponse(
+    val file: Path,
+    val title: String
+)
